add proper interactor; mvp v1
