+++ app/src/main/java/com/example/android/playlistmaker/player/presentation/PlayerView.kt
@@ -1,0 +1,7 @@
+package com.example.android.playlistmaker.player.presentation
+
+interface PlayerView {
+    fun changePlayButton(isPlay: Boolean)
+
+    fun updateElapsedTime(time: Long)
+}
